keep_agent/requirements.txt
twitter.common.app==0.3.0
twitter.common.collections==0.3.0
twitter.common.contextutil==0.3.0
twitter.common.dirutil==0.3.0
twitter.common.lang==0.3.0
twitter.common.log==0.3.0
twitter.common.options==0.3.0
twitter.common.process==0.3.0
twitter.common.string==0.3.0
twitter.common.util==0.3.0
